+++ features/step_definitions/wallets-steps.js
@@ -78,7 +78,7 @@ Given(/^I see the add wallet dialog$/, function () {
   return addWalletDialog.waitForDialog(this.client);
 });
 
-Given(/^I see delete wallet dialog$/, function () {
+Given(/^I see Rename Wallet dialog$/, function () {
   return this.client.waitForVisible('.RenameWalletConfirmationDialog_dialog');
 });
 
@@ -299,7 +299,7 @@ When(/^I submit the create wallet recovery phrase entry dialog$/, function () {
   return this.waitAndClick('.WalletRecoveryPhraseEntryDialog .primary');
 });
 
-When(/^I click on delete wallet button$/, async function () {
+When(/^I click on Rename Wallet button$/, async function () {
   return this.client.click('.RenameWalletButton_button');
 });
 
@@ -311,7 +311,7 @@ When(/^I click on the "Make sure you have access to backup before continuing" ch
   return this.waitAndClick('.RenameWalletConfirmationDialog_dialog .SimpleCheckbox_root');
 });
 
-When(/^I submit the delete wallet dialog$/, function () {
+When(/^I submit the Rename Wallet dialog$/, function () {
   return this.client.click('.RenameWalletConfirmationDialog_dialog .primary');
 });
 
@@ -335,7 +335,7 @@ Then(/^I should not see the create wallet recovery phrase entry dialog anymore$/
   return this.client.waitForVisible('.WalletRecoveryPhraseEntryDialog', null, true);
 });
 
-Then(/^I should not see the delete wallet dialog anymore$/, function () {
+Then(/^I should not see the Rename Wallet dialog anymore$/, function () {
   return this.client.waitForVisible('.RenameWalletConfirmationDialog_dialog', null, true);
 });